Move small utility functions to "utils"
